refactor(scripting): correct the cache-scope retention comment and isolate its test
The LoadContextScopes doc claimed a superseded context's cache entries are
"stranded" and the context collected. That is wrong: _typeCache holds
CompiledScript.Context strongly for the singleton's lifetime, so a
superseded helper context and every assembly loaded into it are retained
for the process lifetime instead. Corrected the comment to say so, and
noted _typeCache as what pins it.

Also: removed a comment at GenerateCacheKey's |alc: append that duplicated
CompileToInstanceAsync's, trimmed the CreateFromImage/null-Display root
cause to its one home (GetCacheScope's doc) instead of three, collapsed the
scope id format to alc{id} (dropping the unobserved Name-based diagnostic
claim, keeping the load-bearing incrementing id), and added a note on
GetCacheScope explaining why ConditionalWeakTable's factory re-entrancy is
expected and must not be "fixed" into TryGetValue + Add.

Moved ScriptEngine_Compiles_Same_Mapping_Against_Different_Helper_Sets_
Without_Cross_Contamination, IHelperValueMapping, and BuildHelperMapping
out of SandboxedScriptingTests.cs (whose doc says its tests run without a
DI container) into a new ScriptEngineHelperSetIsolationTests.cs.
